+++ src/components/calcite-block-section/calcite-block-section.tsx
@@ -8,6 +8,9 @@ import { guid } from "../utils/guid";
 import classnames from "classnames";
 import { CalciteBlockSectionToggleDisplay } from "../interfaces";
 
+/**
+ * @slot - A slot for adding content to the block section.
+ */
 @Component({
   tag: "calcite-block-section",
   styleUrl: "calcite-block-section.scss",
@@ -132,7 +135,7 @@ export class CalciteBlockSection {
       );
 
     return (
-      <section aria-expanded={open ? "true" : "false"}>
+      <section aria-expanded={open.toString()}>
         {headerNode}
         <div class={CSS.content} hidden={!open}>
           <slot />