+++ src/components/calcite-block/calcite-block.tsx
@@ -8,6 +8,7 @@ import CalciteScrim from "../utils/CalciteScrim";
 /**
  * @slot icon - A slot for adding a trailing header icon.
  * @slot control - A slot for adding a single HTML input element in a header.
+ * @slot - A slot for adding content to the block.
  */
 @Component({
   tag: "calcite-block",
@@ -140,7 +141,7 @@ export class CalciteBlock {
     ) : !disabled ? (
       <slot name={SLOTS.control} />
     ) : null;
-    const hasIcon = this.el.querySelector("[slot=icon]");
+    const hasIcon = el.querySelector(`[slot=${SLOTS.icon}]`);
     const headerContent = (
       <header class={CSS.header}>
         {hasIcon ? (
@@ -156,14 +157,8 @@ export class CalciteBlock {
       </header>
     );
 
-    let hasControl = false;
-    let hasContent = false;
-
-    for (let i = 0; i < el.children.length; i++) {
-      const isControl = el.children[i].slot === SLOTS.control;
-      hasControl = hasControl || isControl;
-      hasContent = hasContent || !isControl;
-    }
+    const controlSlot = el.querySelector(`[slot=${SLOTS.control}]`);
+    const hasContent = el.children.length > (controlSlot ? 1 : 0);
 
     const headerNode = (
       <div class={CSS.headerContainer}>
@@ -175,7 +170,7 @@ export class CalciteBlock {
             title={toggleLabel}
           >
             {headerContent}
-            {hasControl ? null : (
+            {controlSlot ? null : (
               <CalciteIcon
                 size="16"
                 path={open ? chevronUp16 : chevronDown16}
@@ -191,7 +186,7 @@ export class CalciteBlock {
 
     return (
       <Host>
-        <article aria-expanded={collapsible ? (open ? "true" : "false") : null} aria-busy={loading}>
+        <article aria-expanded={collapsible ? open.toString() : null} aria-busy={loading}>
           {headerNode}
           <div class={CSS.content} hidden={!hasContent || !open}>
             <CalciteScrim loading={loading} disabled={disabled}>