+++ src/components/calcite-tip-manager/calcite-tip-manager.tsx
@@ -17,6 +17,9 @@ import CalciteIcon from "../utils/CalciteIcon";
 import { getElementDir } from "../utils/dom";
 import { CalciteTheme } from "../interfaces";
 
+/**
+ * @slot - A slot for adding `calcite-tips`.
+ */
 @Component({
   tag: "calcite-tip-manager",
   styleUrl: "./calcite-tip-manager.scss",
@@ -272,7 +275,7 @@ export class CalciteTipManager {
         <div
           class={CSS.container}
           hidden={closed}
-          aria-hidden={closed}
+          aria-hidden={closed.toString()}
           tabIndex={0}
           onKeyUp={this.tipManagerKeyUpHandler}
           ref={this.storeContainerRef}